requirements.txt
numpy
django-common-helpers
python-dateutil
djangorestframework
jinja2
django-ckeditor
django~=3.1.7
django-crispy-forms
celery
redis
python-decouple
django-rq
django_cron
django_crontab
django-celery-results
celery-progress
pandas
# python -m pip install --upgrade Pillow >>> install pillow with this

# MarkupSafe	2.1.1	2.1.2
# Pillow	9.3.0	9.4.0
# PyJWT	2.5.0	2.6.0
# amqp	5.1.1	5.1.1
# appdirs	1.4.4	1.4.4
# asgiref	3.5.2	3.6.0
# async-timeout	4.0.2	4.0.2
# backports.zoneinfo	0.2.1	0.2.1
# billiard	3.6.4.0	4.1.0
# celery	5.2.7	5.2.7
# celery-progress	0.1.3	0.1.3
# certifi	2022.9.24	2022.12.7
# cffi	1.15.1	1.15.1
# charset-normalizer	2.1.1	3.0.1
# click	8.1.3	8.1.3
# click-didyoumean	0.3.0	0.3.0
# click-plugins	1.1.1	1.1.1
# click-repl	0.2.0	0.2.0
# colorama	0.4.5	0.4.6
# cryptography	38.0.1	39.0.1
# defusedxml	0.7.1	0.7.1
# dj-rest-auth	2.2.5	3.0.0
# django-allauth	0.52.0	0.52.0
# django-celery-results	2.4.0	2.4.0
# django-ckeditor	6.3.2	6.5.1
# django-common	0.1.51	0.1.51
# django-common-helpers	0.9.2	0.9.2
# django-cors-headers	3.13.0	3.13.0
# django-crispy-forms	1.14.0	2.0
# django-cron	0.6.0	0.6.0
# django-crontab	0.7.1	0.7.1
# django-downloadview	2.3.0	2.3.0
# django-environ	0.9.0	0.9.0
# django-filter	22.1	22.1
# django-js-asset	2.0.0	2.0.0
# django-picklefield	3.1	3.1
# django-resized	1.0.2	1.0.2
# django-rq	2.5.1	2.7.0
# django-utils-six	2.0	2.0
# djangorestframework	3.14.0
# elevation	1.1.3
# fasteners	0.18
# googlemaps	4.6.0
# idna	3.4
# kombu	5.2.4
# numpy	1.23.4
# oauthlib	3.2.1
# packaging	21.3
# pandas	1.5.2
# pip	22.2.2
# prompt-toolkit	3.0.36
# pycparser	2.21
# pyparsing	3.0.9
# pyproj	3.4.0
# python-dateutil	2.8.2
# python-decouple	3.6
# python3-openid	3.2.0
# pytz	2022.4
# query	0.0.1
# redis	4.3.4
# requests	2.28.1
# requests-oauthlib	1.3.1
# rq	1.11.1
# setuptools	65.3.0
# six	1.16.0
# sqlparse	0.4.3
# types-cryptography	3.3.23
# tzdata	2022.4
# urllib3	1.26.12
# vine	5.0.0
# wcwidth	0.2.6
# wheel	0.37.1
# wrapt	1.14.1